Saving txt2img generation history

+++ app/helpers.ts
@@ -1,3 +1,3 @@
 export function get_random_seed() {
-  return Math.floor(Math.random() * 9007199254740991);
+  return BigInt(Math.floor(Math.random() * 9007199254740991));
 }
